requirements.txt
pandas
numpy
regex
scikit-learn==1.3.0
datetime
cloudpickle
fastapi==0.105.0
uvicorn
pytest
httpx
requests
azure-identity
azure-ai-ml
python-dotenv
xgboost==1.6.2
hyperopt
mlflow
fsspec
azureml-fsspec
azureml-mlflow
argparse
scipy
